requirements.txt
feedparser
feedgen
beautifulsoup4
python-dateutil
requests
tenacity
